requirements.txt
aio-pika==6.5.2
aiormq==3.2.1
idna==2.9
joblib==0.14.1
multidict==4.7.5
nltk==3.4.5
numpy==1.18.1
pamqp==2.3.0
pandas==1.0.1
pyaml==19.12.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sklearn==0.0
yarl==1.4.2
